refactor : #86 core 모듈 리팩토링
