+++ escape-core/src/main/java/com/sparta/domain/user/repository/UserRepository.java
@@ -11,20 +11,21 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
+    List<User> findByUserType(UserType userType);
+    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
+
     default User findByIdOrElseThrow(Long userId) {
         return findById(userId).orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
     }
-
-    List<User> findByUserType(UserType userType);
 
     default User findByEmailOrElseThrow(String email) {
         return findByEmail(email).orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
     }
 
-    Optional<User> findByEmail(String email);
-
-    default User findByUserId(Long userId) {
-        return findById(userId).orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
+    default void throwIfEmailExists(String email) {
+        if (existsByEmail(email)) {
+            throw new UserException(UserErrorCode.USER_DUPLICATION);
+        }
     }
-
 }
