+++ escape-core/src/main/java/com/sparta/domain/user/service/AuthService.java
@@ -4,8 +4,8 @@ import com.sparta.domain.user.entity.User;
 import com.sparta.domain.user.repository.UserRepository;
 import com.sparta.global.exception.customException.AuthException;
 import com.sparta.global.exception.errorCode.AuthErrorCode;
-import com.sparta.jwt.JwtProvider;
-import com.sparta.jwt.RefreshTokenService;
+import com.sparta.global.jwt.JwtProvider;
+import com.sparta.global.jwt.RefreshTokenService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
