+++ escape-core/src/main/java/com/sparta/global/security/filter/CustomAccessDeniedHandler.java
@@ -1,7 +1,7 @@
-package com.sparta.security.filter;
+package com.sparta.global.security.filter;
 
 import com.sparta.global.exception.errorCode.SecurityErrorCode;
-import com.sparta.security.ResponseUtil;
+import com.sparta.global.util.ResponseUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
